+++ engine/flink/management/dev-model/src/main/scala/pl/touk/nussknacker/engine/management/sample/DevProcessConfigCreator.scala
@@ -196,10 +196,11 @@ class DevProcessConfigCreator extends ProcessConfigCreator {
                   None,
                   Some(FixedValuesParameterEditor(List(FixedExpressionValue("'test'", "test")))),
                   None,
+                  None,
                   None
                 ),
-                "bar" -> ParameterConfig(None, Some(StringParameterEditor), None, None),
-                "baz" -> ParameterConfig(None, Some(StringParameterEditor), None, None)
+                "bar" -> ParameterConfig(None, Some(StringParameterEditor), None, None, None),
+                "baz" -> ParameterConfig(None, Some(StringParameterEditor), None, None, None)
               )
             )
           )
@@ -209,7 +210,7 @@ class DevProcessConfigCreator extends ProcessConfigCreator {
           SingleComponentConfig.zero.copy(
             params = Some(
               Map(
-                "bar" -> ParameterConfig(Some("'barValueFromProviderCode'"), None, None, None)
+                "bar" -> ParameterConfig(Some("'barValueFromProviderCode'"), None, None, None, None)
               )
             )
           )
@@ -232,12 +233,14 @@ class DevProcessConfigCreator extends ProcessConfigCreator {
                   None,
                   None,
                   Some(List(MandatoryParameterValidator)),
+                  None,
                   None
                 ),
                 "overriddenByFileConfigParam" -> ParameterConfig(
                   None,
                   None,
                   Some(List(MandatoryParameterValidator)),
+                  None,
                   None
                 )
               )