+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/compile/nodecompilation/NodeDataValidator.scala
@@ -69,8 +69,8 @@ class NodeDataValidator(modelData: ModelData, fragmentResolver: FragmentResolver
         case a: Join => toValidationResponse(compiler.compileCustomNodeObject(a, Right(branchContexts), ending = false))
         case a: CustomNode =>
           toValidationResponse(compiler.compileCustomNodeObject(a, Left(validationContext), ending = false))
-        case a: Source => toValidationResponse(compiler.compileSource(a))
-        case a: Sink   => toValidationResponse(compiler.compileSink(a, validationContext))
+        case a: SourceNodeData => toValidationResponse(compiler.compileSource(a))
+        case a: Sink           => toValidationResponse(compiler.compileSink(a, validationContext))
         case a: Enricher =>
           toValidationResponse(
             compiler.compileEnricher(a, validationContext, outputVar = Some(OutputVar.enricher(a.output)))
@@ -106,7 +106,7 @@ class NodeDataValidator(modelData: ModelData, fragmentResolver: FragmentResolver
             )
           )
         case a: FragmentInput => validateFragment(validationContext, outgoingEdges, compiler, a)
-        case Split(_, _) | FragmentUsageOutput(_, _, _, _) | FragmentInputDefinition(_, _, _) | BranchEndData(_) =>
+        case Split(_, _) | FragmentUsageOutput(_, _, _, _) | BranchEndData(_) =>
           ValidationNotPerformed
       }